Minor changes on the plugin system made.
Major changes on the SpeechGrammarBuilder made.
Added a way to export and load the created grammar to and from xml.
I made it with Telerik RadControls for WinForms.

Notes: I've added an .xml grammar in the TestPlugin for the testing...

SpeechEngine still in the SpeechGrammarBuilder just for the testing.

Next tasks:
1. To figure out a way to make VisualStudio solution file programatically... Oh, boy.
2. To find a way to "copy" the speech recognition setting from one computer to another.
 -> maybe with a server - everyone has an account and we store there this information and the pluggins? This way everyone could use the program with his specific settings almost everywhere.
3. To make the MainForm for the application and connect the rest forms to it. Make it beauty.

After these tasks:
1. To add a way to execute the programs language through C#? Ico?
2. To write whole bunch of plugins.

Others:
1. To buy a microphone - i'll do this... a professional one, because with mine the quality and level of the sound are not so high and the recognition endinge strugles to find out what did i say... especially with my accent :D
2. Consider the possibility to find someone with an good accent for the presentation...
3. Dude, I figured out a way to add custom words into the SpeechEngine dictionary... Cool, ah?

Important:
Dude. It's going to be a huge thing if we keep working hard.
What do you think of the possibility to make a 'market' for plugins, where almost everyone can upload his plugin (after we allow him) and then the users buy it... We could get smth like 5-10%? Something like AppleStore.